Changed flow to splash screen-> if logged open home fragment else open welcome fragment
