+++ app/src/main/java/com/example/coronadiagnosticapp/ui/fragments/splash/SplashViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.coronadiagnosticapp.ui.fragments.welcome
+package com.example.coronadiagnosticapp.ui.fragments.splash
 
 import androidx.lifecycle.ViewModel
 import com.example.coronadiagnosticapp.data.repository.Repository
 import javax.inject.Inject
 
-class WelcomeViewModel @Inject constructor(val repository: Repository) : ViewModel() {
+class SplashViewModel @Inject constructor(val repository: Repository) : ViewModel() {
     fun isLoggedIn() = repository.isLoggedIn()
     fun setIsFirstTime(isFirstTime: Boolean) = repository.setIsFirstTime(isFirstTime)
 }
