+++ app/src/main/res/values/strings.xml
@@ -94,4 +94,7 @@
         ]]>
     </string>
     <string name="city_back_on_tracks">Together we\'ll get the city back on tracks.</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
